+++ RELEASENOTES.md
@@ -1,5 +1,15 @@
 # Release Notes
 
+## What is New in PSAuth Unreleased
+
+August 08, 2019
+
+- Change `psakefile.ps1` to detect Azure Pipelines correctly.
+- Added PowerShell Gallery badge to `README.md`.
+- Update URL Encoding to use more modern URL encoding method
+  `[System.Web.Security.AntiXss.AntiXssEncoder]::UrlEncode()` - fixes
+  [Issue #7](https://github.com/PlagueHO/PSAuth/issues/7).
+
 ## What is New in PSAuth 0.1.2.36
 
 August 06, 2019